[go-gen] Move StringWrap to StringUtils
